# container_http/backend-api-contract.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "PlantGameBackend",
        "version": "1.0.0",
        "description": "Connecting Kubernetes Container with Firebase"
    },
    "paths": {
        "/guess": {
            "post": {
                "requestBody": {

                },
                "parameters": [
                    {
                        "name": "gameId",
                        "description": "",
                        "schema": {

                        },
                        "in": "header"
                    },
                    {
                        "name": "imageUri",
                        "description": "",
                        "schema": {

                        },
                        "in": "header"
                    },
                    {
                        "name": "guessedSpecies",
                        "description": "This is the latin name of the species the user thinks it is.",
                        "schema": {

                        },
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The guess is going to be processed."
                    }
                },
                "description": "Is called every time the user makes a guess."
            }
        },
        "/parks": {
            "get": {
                "parameters": [
                    {
                        "name": "park",
                        "description": "",
                        "schema": {

                        },
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        },
                        "description": "The list of 5 species inside a plant is given."
                    }
                },
                "summary": "get species of park",
                "description": "This one is called every time the user chooses a park."
            },
            "post": {
                "requestBody": {

                },
                "parameters": [
                    {
                        "name": "parkName",
                        "description": "",
                        "schema": {

                        },
                        "in": "header"
                    },
                    {
                        "name": "speciesList",
                        "description": "This is a list of all species appearing in the specified park.",
                        "schema": {

                        },
                        "in": "header"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "The park and its species is added to the database"
                    }
                },
                "summary": "adding park",
                "description": "It adds a park with a list of species to the postgre database."
            }
        }
    }
}